Eliminate some technical debt
The feature combination #+(and linkage-table (not sb-dynamic-core))
has no reason for being.

The next batch of changes will search-and-replace all #+sb-dynamic-core
occurrences with #+linkage-table. However, in preparation, this slightly
nontrivial change tries to reduce some boolean expressions under an
assumption that the features are truly equivalent.
The old expressions were so hideous that a bug may have crept in to the
new renderings, which is why this is submitted for separate review.