Assume #+(and linkage-table sb-dynamic-core) in tests
At this point the conditionals in source code exist only so that Alpha and HPPA
can build, in theory. I think they don't actually run. If they do, they need to
be updated to conform with the rest of the ports.  As such, the test suite must
not try to circumvent failures on #-linkage-table, but instead, failure should
be seen as a sign to fix the incomplete implementations.