Remove special cases in os_link_runtime
This removes the logic in os_link_runtime which copied 1 linkage entry to static
space below NIL. Instead, the whole linkage space is below NIL, which makes some
entries accessible relative to NULL-TN. All machine-specific code is confined to
a machine-specific file. e.g. some of the architectures need to write the 0th
entry as a function pointer and not a small stub of code.

Also: On MIPS, we were allocating 16 megabytes for linkage entries which is
not only a totally insane size, but all the more because each entry is 4 bytes.
Reduce the table to 64KB total.