Rename and rearrange pseudo-atomic.h
The pseudo-atomic header was more often #included for get_alloc_pointer() than
the p-a accessors,so rename it to 'getallocptr.h', but keep the accessor for
thread->pseudo_atomic_bits in a header that is still named 'pseudo-atomic.h'.

Factor the three major variations into different files. Prior to this change,
too much thinking was needed to figure out which bundle of #defines and #ifs
were applicable to a particular arch x {cheneygc,gencgc} x {+thread,-thread}.