riscv: Shave two instructions on average off gengc allocation.
We want to find a good tradeoff between speed and space for
allocation. PPC emphasizes space and uses traps, while other
architectures use assembly routines. We use an assembly routine, so
move some code out of the inlined portion of the allocation routine
into specialized stubs. Now the inlined portion which sets up the
call-out is always 4 instructions, rather than 5 in the variable-sized
case, 6 in the small constant size case, or 7 in the larger constant
size case.

The speed impact of having an extra call indirection is negligible,
given that this is the slow case. It should be much faster than a
context switch in any case.

This cuts a bit more than 160kb of code out the cold core.