x86-64: Remove clearing of direction flag
If it has to be clear on entry to every function, and we never set it, then
we needn't clear it. For sake of argument, there is not a single use of either
the STD or CLD instruction in the entirety of src/runtime/sbcl as compiled by
either gcc or clang. I'm assuming the same is true of win32.