x86: Fix huge &rest list allocation
This change is the bare minimum to get the test to pass. The assembly routines
are a real morass, and I started to tidy them up more, but lost ambition.

Known issues: the routine for out-of-line allocation should try a pointer bump
before calling into C. No wonder it's so dang slow.
And I can't help but feel that the x87 fpu state might be clobbered.