Remove layout alignment requirement
Layouts need to be variable-length to contain a trailing bitmap.
This change works for either word size, but is contrary to using alignment
to make layout pointers representable in 24 bits on 32-bit machines.
I still think that there is benefit to aligning layouts at a grain which
allows carrying a "funny" pointer in each instance's header data and
eliminating the LAYOUT slot, but I have no imminent plan to do so.