riscv: Fix potential interrupt safety and alignment issues.
The stub was leaving the number stack unprotected. I was concerned
about this but then convinced myself it wasn't an issue because of
pseudo-atomic. This is actually not true because pseudo-atomic does
not protect the C stack from interrupts trashing the unprotected stack
area.

So pass the argument to alloc-tramp on the control stack instead. This
also avoids the number stack alignment issue, which is great because
we were not computing the aligned size properly after the introduction
of alloc-tramp stubs.