ppc64: Rearrange pointer lowtags
This shaves off an instruction in the CAR, CDR, and %INSTANCE-REF vops
and corresponding setters. NIL as a list is perfectly fine, but as a
symbol it requires more care to access the slots because the difference
between other-pointer and list-pointer is not N-WORD-BYTES.

Other than that, I'm surprised at how well our abstractions worked.