Remind genesis how to cold load non-fopcompiled functions.
Reintroduce FOP-FSET, and dump it only when cross-compiling block
compiled DEFUNs, as it is not possible to mix block compilation with
FOP compilation. This also makes it easy to allow cross-compile to
link functions from ordinary compilation again (i.e. not from
fop-compilation) in the future if we so choose.

This does a more direct port of what CMUCL does (which incidentally
does cold load some block-compiled functions), rather than the
pre-2015 code which probably didn't work with block compilation and
was a little messier anyway.