Decouple IR1 from IR2.
It may seem somewhat simplistic to just eliminate the type constraint,
but on inspecting the CMU CL sources closer, one can tell very clearly
that IR1 is intended to not to depend on details of IR2. As proof, one
sees that none of the other info slots of the IR1 data structures have
any reference to IR2 backend structures. Indeed, they are just named
"info", and commented as a slot used by the backend, the details of
which IR1 need not concern itself.

For the sake of consistency with everything else and abstraction
purposes it seems better to return this slot to its primordial state
and delete the annotations. Plus, it gets rid of a kludge and a
def!struct. The type test was used as an assertion on the possible IR2
values it has, which, again, for abstraction purposes, should belong
to the backend rather than the frontend.