x86-64: Optimize out a widetag test in PCL-INSTANCE-P
Since #+compact-instance-header causes all functions to have a layout in the
header, we can do one fewer check when reading a funcallable-instance's layout.

Weakening the arg type on %FUNCALLABLE-INSTANCE-LAYOUT didn't make sense to me,
though I thought about it. The new accessor %FUN-LAYOUT implies more generality,
and is far easier to type, so that settled it.
I think we never need to signal an error if something is a function but not a
funcallable-instance, which would have contraindicated this approach.