Rename LAYOUT-FOR-STD-CLASS-P -> LAYOUT-FOR-PCL-OBJ-P
Slight fix to a naming problem stemming from rev 62f25b3b18 and 1913e44aa3
and surviving through various alterations, which is that STD-INSTANCE-P
does NOT include funcallable objects.

But now we have another problem in that
  ;;; The class STD-CLASS is an implementation-specific common
  ;;; superclass of the classes STANDARD-CLASS and FUNCALLABLE-STANDARD-CLASS
but again STD-INSTANCE is the counterpart to FSC-INSTANCE, and an instance
is exclusively one or the other, so it is bad that STD-CLASS implies a union.