Test for noop dlclose implementations
POSIX does not require dlclose(3) to remove objects from the address
space. While many libcs perform reference counting and remove objects, not all
libcs do (see: msul libc).

Update documentation to state this is a possibility. Additionally, check if
dlclose is a noop while running regression tests and skip all tests that depend
on objects being unmapped if it is.