Substitute dummy stream for stream-error involving dynamic-extent
This helps with situations where a signaled condition outlives the
dynamic extent of a stream that the condition references, and then
the user tries to print the condition.
On the other hand, this loses information if the user would have used
the condition while still within the stream's extent.
It's hard to know what to do. Keeping any more state would just add
overhead that dynamic-extent was trying to avoid.