Mostly remove the compiled-debug-fun hash-table from SB-DI
We can poke values into our own objects, so a side table is overkill.
Just store the debug fun back into the object to which it pertains.
On cheneygc, do what we always did because code may be immutable.
Also, delete all the memoized debugger debug funs when saving a core.

Incidentally the comment about unbounded table growth was wrong.
It was made a weak table 11 years ago in rev 9abfd1a2b2.