Make 'print' and 'reader' amenable to slam.sh
We need not to lose the specialization recorded for global constant arrays.
There's no way to know in MAKE-ARRAY that it will be stored in a defconstant,
so just force the issue by adding a keyword. It's not pretty, but unless
somebody is extremely motivated to attempt the other approach outlined in
specializable-array, it'll have to do.
We bit the bullet on CL:FLOAT etc, so it might be no more tedious to shadow
CL:ARRAY with SB-XC:ARRAY as a structure that carries around one extra piece
of information. It won't be fun, and won't have a high payoff - all it would
achieve is removal of the new keyword arg from SB-XC:MAKE-ARRAY.