Remove nearly all address-based hashing from system hash-tables
Most of them could be made to use stable hash values.
A few remain as possibly hashing by address:
- EQL-specializers (depending on what you put in it)
- finalizers (obviously)
- unknown-type atoms (aren't these symbols ?)
- allocation-point fixups (probably doesn't need to be a hash-table)

Perhaps it's ironic that an EQL table is cheaper than an EQ table if all keys
are symbols. The memory consumption is the same, and the EQL table will never
rehash due to object movement.
Perhaps we should use SYMBOL-HASH for symbols stored in EQ tables too, so that
users can have this benefit without having to know this "One Weird Trick"