Reduce hash-table size to 7 if unspecified
One extra doubling to attain the former initial size of 14 does not affect
asymptotic behavior in the least.  Using make-host-1 as a measure of "typical"
table usage, 10% of all tables never got a single key inserted, 20% got exactly
1 key, and maximum usage of 0 through 7 keys encompassed 75% of all tables.

The memory consumption (for 64-bit cpu) of a default table is now 448 bytes:
* (sb-aprof:aprof-run #'make-hash-table)
       %        Bytes        Count    Function
 -------  -----------    ---------    --------
  57.1            256                 %MAKE-HASH-TABLE
      62.5          160          1        SIMPLE-VECTOR
      37.5           96          2        (SIMPLE-ARRAY (UNSIGNED-BYTE 32) (*))
  42.9            192            1    %ALLOC-HASH-TABLE - HASH-TABLE
 =======  ===========
 100.0            448

Prior to git rev b16db3ecdc (the first of many changes that affected
hash tables) it was in excess of 1200 bytes.