Rename *EXIT-IN-PROCESS* -> *EXIT-IN-PROGRESS*
To quote https://writingexplained.org/in-process-vs-in-progress-difference
"'in progress' has been the preferred version ... for nearly 200 years".

In addition, "process" has an overloaded connotation - like maybe this
variable served to distinguish thread vs process exit. Does it? If it does,
then I really don't get it. Was it purposeful double entendre? Yikes.