Sort of work around Darwin problem in fcb-threads
Try both sigwait() and sigsuspend(), the latter of which delivers the signal.
So based on a new thread-local hack flag, the signal is ignored.
Now the test usually prints a note that it applied that workaround, but instead
it might hang.  I have a feeling that the OS issue is that the signal becomes
unpending by magic before the sigsuspend().

Also add a less stressful "smoke test" that runs without a background GCing
thread. This test worked for me on win32 as well.