Fix race with GC in new_thread_trampoline
thread->control_stack_end must be strictly above any value that SP could have
for the few instructions surrounding funcall1(). In the x86-64 call convention,
'arg' is not necessarily above all the local vars, because it is passed in a
register and stored wherever the compiler wants.  Adding 1 word ensures that
the value is high enough to prevent the "no SP known for thread" error.
I considered using __builtin_frame_address, but the more I thought about it,
the more I concluded that the current approach is valid.