Reduce exposed surface area of SB-UNIX
Shake out the internal symbols. I'm sure there will be complaints,
but it's essential to find out what people are (inadvisably) using.

The comment about what is used "only in this file" was not quite true.
Inlining requires saving source code, and the theoretical possibility
exists of some other system file needing the INT-SYSCALL macro (etc)
if deciding to inline UNIX-READ or similar.