Remove some libc wrappers except on netbsd
We have "modulating" wrappers because of structure arguments of unknown size,
do-nothing wrappers because of linker and preprocessor tricks, wrappers because
libc functions are allegedly buggy, and wrappers because ldso-stubs was a thing.
Enough is enough. Don't wrap if we don't need to, so that at least some clarity
is gained in the case where we do need to, and so that someone (such as myself)
doesn't think we can remove sb_gettimeofday (since it looks useless) only to
find through an archaeological expedition that it might be for netbsd.

Conditionally using a wrapper, as for the transcendental math functions, is
preferable to introducing mystery wrappers. My only complaint is that we
have 39 different ways of deciding to vary the name of the alien function.
That could at least be unified, given sufficient motivation.