64-bit: Use microseconds for internal-time-units-per-second
Move the first clock_gettime as soon as possible, because the process acrues
CPU time in main(), and with the finer resolution it is often observed to have
more runtime than realtime on startup which does not make a lot of sense with
only one thread. And I had forgotten to initialize 'lisp_init_time' on SunOS,
so it's better to do it in exactly one place.