Remove wallclock time from SB-SPROF sampling
To be compatible with pure :sb-safepoint (i.e. not the combo of safepoint
plus pseudo-atomic), the data recorder needs to be made async-signal-safe
which means rewriting it into C. So it can't use GET-INTERNAL-REAL-TIME,
and while it might use the signal-safe clock_gettime call directly,
we should not be in the business of gathering data for purposes other than
SB-SPROF:REPORT which just ignores the wallclock time.
(The SPROF API was always marked "Experimental, subject to change")

And as things are, typically at least 3% of all CPU time in a profiling
run is spent in the profiler itself, not to mention its massive consing
overhead. So it really warrants a rewrite into C for good reasons.