Refactor layout bitmap bit extractor
There are two ways of iterating over descriptor slots in C, either a test
of "logbitp" at each word index, or the special code in scav_instance() which
avoids a div and mod operations at each step. But the logbitp approach had to
decide at each iteration whether the bitmask was a fixnum or a bignum.
This change removes that decision by canonicalizing the bitmask once up front,
and further anticipates a change in which the bitmask is moved to the end
of the layout so that we never have to chase a pointer outside of the layout
to utilize the layout, which results in generally more understandable C code.

In fact, I don't see how the assertion about a bignum for the bitmask in
coreparse heap relocation was correct, because it used layout->bitmap
without adjustment. But then it used the adjusted value when iterating.
So the code was right, but the assertion probably would fail.
That's one example of why an external bignum is problematic.