Fix editcore
Misuse of /WORD versus /SIGNED-WORD for layout bitmap extraction resulted in
the wrong answer from INTEGER-LENGTH, which caused a classic buffer overrun,
scanning more words of the current object than it actually has.
But miraculously it still produced correct ELF files, because writing a fixup
clears the source word, which makes it not get a fixup when visited as part
of the object that properly contains it. So unfortunately the self-correcting
nature makes it nearly impossible to assert validity of the output,
at least if the tool is treated as an opaque box. I'm not sure what to do.