Add assembly code quality test for TYPEP transforms
For better or for worse, this uses a golden file which is unduly difficult
to produce because of the various build passes and inability to read
floating-point numbers in their normal syntax during self-build.
Warm build writes a file that it can't re-read in dump-and-load,
as type names occur in the file that aren't known until later, but
the file is sorted alphabetically to maintain a well-defined order.
The data are collected under control of some #+nil'ed out code in the
typep source-transform plus hand editing to work around problems.

Using this technique I was able to see that a small change in widetags
could remove 2 instructions from SIMPLE-UNBOXED-ARRAY without ill effect.
That will be checked in subsequently.