Checkin some debugging code
This is stuff I keep in my local tree. It augments the overwhelming
variety of "show" functions, but fortunately, this one is consistently
not harmful and almost always helpful. All the others force context
switches and pretty much ruin any hope of figuring out whatever it is
that precipitated the need for such manner of debugging to begin with.