x86-64: Remove uses of reg-in-size from test file
(except for one test that asserts that it inhibits optimizations)

Though REG-IN-SIZE looks almost exactly like SIZED-THING, the former is
harmful because any vop generator that uses it causes machine instructions
to receive operands that can not be related back to the generating vop,
so we can't do any dataflow analysis on the abstract assembly code.

In contrast, SIZED-THING is used only within an instruction encoder
to inform the prefix emitter which operand size prefix to insert,
and has nothing calling it outside of the 'insts' file.