Remove a hack from echo-stream-misc
ANSI-STREAM-PEEK-CHAR is special-cased to understand echo-streams,
but now with 1 less problem - that of the unique argument convention.
No other builtin stream type would ever have been able to imnvoke the
misc method on :PEEK-CHAR because *only* CL:PEEK-CHAR knew how to stuff
3 args into 2. (And why like that? Why not a DX list of 3 values in arg1?)
Well it didn't matter, because :PEEK-CHAR is not an operation that is ever
forwarded from one stream to another.

Having done this we can remove the second optional argument to all the
misc entry points because it's not used for anything any more.