Add a stream flag for simple-stream
Using a flag bit, it should be possible to adapt all of the CL stream
APIs to dispatch on the other stream family without having to redefine
a ton of CL: functions. So we can anticipate that the simple-streams
contrib might be loaded, but not define the support unless it is.
And then only redefine 4 of so functions that add extra arguments.
Anyway, proceeding in baby steps for now.