Fix an inefficiency in encapsulating streams
When forwarding an operation from an SBCL-builtin ("ANSI") stream that
wraps a CLOS stream, dispatch to one of two groups of functions.

For a standard function CL:mumble, the Gray streams should (usually)
defer to a function named STREAM-mumble. The simple-streams on the other hand
use something else, however, we don't have an easy handle on that something
(yet). So simple-streams have to go the roundabout way all the way back to
the CL function which then has to re-decide that it has a simple-stream,
in adition to performing the redundant check for T or NIL as the stream.

Hence the advantage for Gray streams after this is that we skip the circuitous
routing through the CL API layer.