Start the finalizer thread early and often
This change should resolve both lp#1906571 and lp#1907872.

Delaying finalizer thread creation until just-in-time was too complicated.
Doing it at startup is less confusing because the state of the world is known,
and synchronously stopping/restarting around any call to FORK, as well as
immediately prior to SAVE-LISP-AND-DIE, is easier too.

Use a native condition variable to inform the thread when to run or exit.
Deadlock can't happen because triggering finalizers never involves make-thread
or a mutex (though native condition vars may use one internally), and there's
no hidden "possibly starting a thread" state.