Distinguish between readers and writers in INVOKE-EMF
The heuristic previously used, based on the number of arguments passed,
fails in the case that code calls the generic function with the wrong
number of arguments, gets a cache miss, and then scribbles over the memory
pointed to by the second argument.

INVOKE-EFFECTIVE-METHOD-FUNCTION does not need this disambiguation,
because it receives information about the number of required arguments
of the generic function called, not just the number of arguments at
the call site.

Fixes lp#1909659