Try using a futex for the finalizer wake flag
I have a feeling that our use of a signal for stop-for-GC causes problems
with a POSIX condition variable for the finalizer "run" indicator.
I don't see why, but I have an inkling that this might be the problem.

If this change resolves the unexplained test suite failures on Linux,
then I'll have to figure out a similar fix for non-futex.
Win32 can almost surely remain as-is, since signals aren't involved.