Simplify FIND-CONSTANT et. al.
* FIND-CONSTANT when compiling to memory was massively overengineered.
  CORE-COALESCE-P made no sense either as a variable name or a concept,
  since the standard expressly disallows coalescing in COMPILE.

* MAYBE-EMIT-MAKE-LOAD-FORMS received a NAME argument for the sole purpose
  of advising the function to do absolutely nothing.

* EMIT-MAKE-LOAD-FORM with a NAME did something totally different, and wrong.
  There should not be any circularity checking performed.

* Named constants are henceforth actual things- the %SOURCE-NAME slot in
  a #<CONSTANT> meaningfully conveys that the object was a defconstant.

Most of the weirdness came from git rev 45bc305be4. This change tried not to
alter behavior, however it did expose a very sketchy allowance for anonymous
constants that are EQ to a named constant where sometimes we permit a structure
lacking a MAKE-LOAD-FORM method to be dumped as long as it is EQ to a
defconstant earlier in the file. The comments above FIND-CONSTANT point out two
such occurrences. I did not fix that. I imagine there is a portability issue as
pertains to non-conforming code that SBCL causes to work, but should not work.