do not hardcode LD=ld in GNUmakefile
When cross compiling for an alternate libc implementation (e.g. musl for
static linking) a different linker required. It is currenly impossible
to pass LD=some-other-linker-ld in the environment or as part of a call
to make because the GNUmakefile hardcodes LD=ld. This commit fixes that
an also make it so that make-target-contrib.sh inhertis LD in the same
way that it inherits CC.