Improve sb-cover's codegen for non-x86
Store the coverage mark bytes in the code header, similarly to what was
done for x86 and x86-64, but use CODE-TN as the base register and NULL-TN
as the value to store. It's usually just 1 instruction per mark now.

Clearly there is much low-hanging fruit such as the following:
; 178: L3:   44A0C5E5         STRB NULL, [CODE, #68]
; 17C:       45A0C5E5         STRB NULL, [CODE, #69]
; 180:       46A0C5E5         STRB NULL, [CODE, #70]
; 184:       47A0C5E5         STRB NULL, [CODE, #71]
; 188:       48A0C5E5         STRB NULL, [CODE, #72]
; 18C:       49A0C5E5         STRB NULL, [CODE, #73]
which could be reduced down to one store with some more analysis.

Ultimately I decided that while this change causes pages of code to be
dirtied from GC's viewpoint, it is no worse than touching pages containing
random conses, and probably better. If anything, a page of mixed conses and
objects may hold as many as (/ gencgc-card-bytes (* 2 n-word-bytes) things
to re-scan in GC, but code pages are never as dense as that.

x86 and x86-64 already improve on both of those by using
software page mark bits and by compressing the mark array.

The best part is that the comment "We're being naughty here" no longer
pertains, and the hack in FIND-CONSTANT is removed.