Remove some weird SB-SPROF external functions
The interface to receiving an individual stack trace (MAP-TRACE-SAMPLES)
should have been a list or vector of locations, and I can conceive of no
reason but for the highly specialized usage in 'disassemble' for wanting
to visit each location that had ever been interrupted (MAP-ALL-SAMPLES).

Had these been good APIs with poorly chosen names (or possibly just bad APIs
with well-chosen names) then  maybe it would have made sense to avoid
breaking them. But they're just horrible- not general enough to expose,
and incredibly confusing given that "samples" means not 1, not 2,
but 3 different things simultaneously: all instances of the DEFSTRUCT,
all traces captured, and all locations in a trace.

Well, they said "EXPERIMENTAL: Interface subject to change." so
here we are- Consider it changed.