Warn about duplicate flet,label,macrolet names
I looked at some other lisps, and there are three different behaviors:
use the first of dup names, or the last, or outright fail to compile.
So this needs to be at least warning-worthy, if not actually an error.
Macrolet used to signal a style-warning, but some people ignore those,
and making all three the same severity seems the most sensible.