Improve sb_toggle_sigprof
If the user had previously asked a thread to block SIGPROF, unmasking it
should work even if the thread receiving the INTERRUPT-THREAD request to
unmask SIGPROF was already handling some other signal.
This change is hard to test, but seems like it should work better this way.

Also revise some comments.