Simplify SB-SPROF:WITH-PROFILING
Don't bind controls passed to START-PROFILING. Nothing uses them afterward.
It might be thought that the intent was to allow nested STOP/START pairs to
see the rebindings, except that START erases all previous samples, so no.

Also there was an "s" missing in "inappropriate set of sampled thread"
and :MAX-SAMPLES was inadequately explained - it's both min and max if looping.