Remove mentions of SIGRTMIN in sa_mask
These came from e6f4c752 with no explanation as to why they might be needed.

One guess is that someone heard that gnu pthreads secretly rob SIGRTMIN and +1
for internal use (as SIGCANCEL and SIGSETXID). If that's what it was, the SBCL
code was flawed, because the C headers hide the kernel's SIGRTMIN, so that
userland SIGRTMIN starts at SIGSETXID+1. i.e. it can't have been a correct way
to block any internal-use signals, if that's what it was.
I don't know what else it may have been aiming at, and it's not worth study
because sb-safepoint is unsupported except on win32, which does not use signals.