x86: emit VECTORP as one instruction for known arrays
Extend vop :GUARD semantics and use it to transform VECTORP to ARRAY-RANK=.

An IR1 transform for this is less workable because selecting the order in which
IR1 transforms run is a problem. A brute-force approach would put the more
specific transform last in source code (because they run in reverse order), but
here the less-specific one is FOLD-TYPE-PREDICATE which should transform VECTORP
to NIL for things that are known non-vectors. So more-specific-last is best.
Yet it seems wrong if not brittle to be so order-sensitive.

Primitive type restrictions aren't enough either, because no union of ptypes
match ARRAY in the language. Letting a vop peek at the call node seems like a good
approach, and it's better than choosing how to emit inside a single vop,
as that would wastefully allocate a temp register needed for the general case.