Simplify genesis structure access utilities
WRITE-SLOTS and READ-SLOTS don't need to be told the type of instance
because metadata can be read from the object itself, and from the descriptor
we can locate the host's description of the target instance type.
The fails for layouts because layout-of-T is created before we have a
descriptor corresponding to layout-of-layout. I could flip the creation
order, but then the inherits vector for layout-of-layout would need to patch
in layout-of-T. One or the other needs extra assistance.

And eliminate layout patching for any purpose other than by CHILL-LAYOUT.
We no longer to need it for classoid-cell, and the only other case was
for MUTEX which can be defined sooner.

Finally, fix a broken print-object method one more time.