Fix edge case in %RANDOM-{SINGLE,DOUBLE}-FLOAT
RANDOM with a float argument used to produce that argument as a return
value with a probability depending on the magnitude of the
argument. That behavior was wrong since the argument is an /exclusive/
upper bound for the range of produced values.

A simple way to demonstrate the problem is

  (loop :repeat 10000 :collect (random least-positive-single-float))

which should evaluate to a list of 0.0f0s, but also contained
LEAST-POSITIVE-SINGLE-FLOAT elements before this change.

Note that the type deriver for RANDOM already treated the argument
correctly so that, for example

  (loop :repeat 10000 :count (= least-positive-single-float
                                (random least-positive-single-float)))

already produced the expected result before this change.