Reinstate and fix "Allow MAYBE-INFER-ITERATION-VAR-TYPE to succeed with multiple SETs"
This reverts commit a4463f9edc171668479dd21cfa42347005f1b04f and fixes
the changes to MAYBE-INFER-ITERATION-VAR-TYPE such that changes to the
iteration variable in different "directions" trigger the conservative
behavior.

Original commit message:

Allow MAYBE-INFER-ITERATION-VAR-TYPE to succeed with multiple SETs

But only if all changes go in the same "direction".

The motivation for this extension was the following loop submitted by
mfiano in #sbcl which had to resort to generic arithmetic for
decrementing the loop counter:

  (loop repeat (bar-num bar)
        for x of-type single-float = 1f0 :then (* x 2f0)
        do (progn))

The reason is that the expansion decrements the loop counter in two
places:

  (LET ((#:LOOP-REPEAT-2224 <known-to-be-of-type-fixnum>) ...)
    (DECLARE ... (TYPE INTEGER #:LOOP-REPEAT-2224))
    (TAGBODY
      (IF (<= #:LOOP-REPEAT-2224 0)
          (GO SB-LOOP::END-LOOP)
          (DECF #:LOOP-REPEAT-2224))
      (SB-LOOP::LOOP-DESETQ X 1.0f0)
     SB-LOOP::NEXT-LOOP
      ...
      (IF (<= #:LOOP-REPEAT-2224 0)
          (GO SB-LOOP::END-LOOP)
          (DECF #:LOOP-REPEAT-2224))
      ...
     SB-LOOP::END-LOOP))