Disallow unexpected warnings of any severity in make-host-1
All architectures are warning-free now.
But of course only trust our own warnings. Other Lisps' warnings are
mostly unhelpful as far as we're concerned. e.g. Clozure warns about
an unused binding in (loop for ((a . b) . c) in x do (frob a b))),
but the workaround is ugly - an effectless (PROGN C) or similar.