Be stricter about lack of errors in cold-init
By making ERROR always print something I saw two hidden problems:
- PCL seems not have all the right compile-time effects to use an
  optimized ctor, which might only be a problem when compiling PCL
  itself, I'm not sure. But it's a failure that is obscured
  due to a handler-bind around an error-producing function.
- Another constant-folding failure, this time on CLOSUREP.

Unfortunately, we can't easily do a swap-out of #'ERROR for a different
function in late warm build, because it has to be inlineable to use
the format-control string tokenizer transform. But NOTINLINE removes
the ability to use a :KNOWN-FUN load-time constant4, which is not big
issue, but really the string transform should be unrelated.