Fix a real bug found by CMU CL build.
CMU CL correctly warned that the cross-compiler's version of CODE-CHAR
can't handle base-char-code-limit, since it is outside the declared
integer ranges. I'm surprised a self-hosted build did not catch this,
but then again, our self-hosted build is not built
with *derive-function-types* set to T, unlike with CMU CL where that
is the default. Perhaps our self-hosted build should be built with
function type derivation on as well.