Rewrite TRANSFORM-MAKE-ARRAY-VECTOR
This change provides a better framework for doing memory allocation using
an allocator that does not assume that heap memory is prezeroed, by
separating allocation from zeroing, and deferring the decision to elide
zeroing until IR2 says it's OK to. It fixes the bug that VECTOR-INITIALIZED-P
failed to recognize some of the ways in which the transform could initialize
the data (and so it would zero-fill and then fill with data.)

Consequently it became possible to elide zeroing of dynamic-extent vectors
whose type is safe for GC without clearing the bits. I rolled that aspect into
this change. If there is any objection, it's simple enough to restore strict
backward-compatibility, though only unportable code should care.