Add a global switch to work around CL pathname nonsense
There are at least two aspects:
 - what pathname should go in *compile-file-pathname*

 - what namestrings should be stored into on-disk artifacts
  (also relates to the namestrings in coverage hash-table)

The spec has nothing to say about the latter, so we can pick anything.
But unfortunately we base the choice on the former, which the spec does say
something about. However, we can clean up a pile of crap and eliminate some
custom handling around this by just making SBCL do the right thing under
control of SB-C::*MERGE-PATHNAMES*. And as it happens, self-build had a use
for this which is now just expoiting the selector switch.