Don't do %define-fun-name-syntax at Genesis time.
Instead of having an abstract-breaking FOP-FUNCALL method, just make
it happen asap in COLD-INIT. Having a cold init method isn't the
prettiest either, but at least it means that we are free to change how
%define-fun-name-syntax is implemented without breaking cold-load and
knowing that cold-init will do the right thing.

This also allows us to remove the GET handler machinery for
fop-funcall and fop-funcall-for-effect.