Rearrange and simplify some stuff.
* info-functions.lisp is gone now, its stuff being split off into more
logical places.
* defmacro bootstrapping is now simpler. The rearrangement allows us
to now do the compile-time part of DEFMACRO as well while
bootstrapping, making the cross-compiler and target-compiler behavior
more similar, also allowing us to avoid a bit of
encapsulation. Avoiding the encapsulation allows us to simplify the
bootstrapping of DEFMACRO and its supporting macros and make it more
coherent. Now the bootstrapping (defining DEFMACRO and friends into
the cross-compiler) happens in macros.lisp where most of the other
macro definitions are defined into the cross-compiler, making it clear
that this doesn't need to happen until right before the cross-compiler
starts running.
* Also remove some superfluous SB-XC: prefixes.