Replace ANSI-idiom with MAYBE-INLINE.
In our own code we should always prefer MAYBE-INLINE since NOTINLINE
inhibits optimization.

From the CMU CL manual, section 5.8.3:

"The problem with using notinline in this way is that in Common Lisp it does more than just suppress inline expansion, it also forbids the compiler to use any knowledge of myfun until a later inline declaration overrides the notinline. This prevents compiler warnings about incorrect calls to the function, and also prevents block compilation."

And not all instances have been changed, because
"When the goal is merely to control whether inline expansion is done
by default, it is preferable to use extensions:maybe-inline rather
than notinline. The notinline declaration should be reserved for those
special occasions when a function may be redefined at run-time, so the
compiler must be told that the obvious definition of a function is not
necessarily the one that will be in effect at the time of the call."