Use a less overloaded term than OFFSET in vops
We had :OFFSET, VECTOR-DATA-OFFSET, :COMPLEX-OFFSET, and who knows what else,
not to mention that "OFFSET" could was both a macro arg and codegen arg in
the same defining form.
There are only 200 grep hits on "ADDEND" (but nearly 10000 on "OFFSET")
so use ADDEND to mean the thing that you add to an INDEX.

DATA-VECTOR-{REF,SET}-WITH-OFFSET still need to be renamed, but I found those
names to be less bothersome than the duplicitous macro arguments.

Make minor corrections to incorrect comments as well.