Rename and improve sb-vm::zero-fill
Zero-fill is now SPLAT, and the former SPLAT is QUICKFILL.
Hopefully this isn't too confusing.

New special cases are handled for splatting :INITIAL-ELEMENT NIL
(hence the rename, because 0 is not NIL) and SB-PCL:+SLOT-UNBOUND+.

This change also anticipates some other cases that will be needed
for further work on #+array-ubsan.