array-ubsan: Allocate shadow bits only if needed
Arrays that have :INITIAL-CONTENTS or -ELEMENT can save space
by not using shadow bits.

A simple and reliable way to decide whether to allocate the bits is by
passing a extra argument to ALLOCATE-VECTOR rather than inferring based
on data flow and seeing or not seeing the initialization step.
(MOVEs can always mess you up if not done carefully)